# expression: implement vectorized evaluation for `builtinIsIPv4MappedSig`