<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for builtinIsIPv4MappedSig , for #12103


### What is changed and how it works?
```
go test -v -benchmem -bench=BenchmarkVectorizedBuiltinMiscellaneousCasesFunc -run=BenchmarkVectorizedBuiltinMiscellaneousCasesFunc -args "builtinIsIPv4MappedSig"
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinMiscellaneousCasesFunc/builtinIsIPv4MappedSig-VecBuiltinFunc-12                 164247              8100 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinMiscellaneousCasesFunc/builtinIsIPv4MappedSig-NonVecBuiltinFunc-12               34106             35027 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinMiscellaneousCasesFunc/builtinIsIPv4MappedSig-VecBuiltinFunc#01-12              177771              7389 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinMiscellaneousCasesFunc/builtinIsIPv4MappedSig-NonVecBuiltinFunc#01-12            34821             35370 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinMiscellaneousCasesFunc/builtinIsIPv4MappedSig-VecBuiltinFunc#02-12              300663              3868 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinMiscellaneousCasesFunc/builtinIsIPv4MappedSig-NonVecBuiltinFunc#02-12            38066             32233 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinMiscellaneousCasesFunc/builtinIsIPv4MappedSig-VecBuiltinFunc#03-12              627260              2127 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinMiscellaneousCasesFunc/builtinIsIPv4MappedSig-NonVecBuiltinFunc#03-12            68324             17375 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      11.416s
```


### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test